requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
pydantic>=2.0.0
tenacity>=8.0.0
anthropic>=0.3.0  # For Anthropic Claude support
openai>=1.0.0     # For OpenAI support
together>=0.2.0   # For Together.ai (Code Llama, DeepSeek)
google-generativeai>=0.3.0  # For Google Gemini models
aiohttp-ratelimiter>=0.1.0  # For rate limiting 